Enum safe_ffi::errors::FfiError

pub enum FfiError {
    CoreError(Box<CoreError>),
    NfsError(Box<NfsError>),
    DnsError(Box<DnsError>),
    PathNotFound,
    InvalidPath,
    PermissionDenied,
    JsonParseError(ParserError),
    JsonDecodeError(DecoderError),
    SpecificParseError(String),
    JsonEncodeError(EncoderError),
    LocalConfigAccessFailed(String),
    Unexpected(String),
    UnsuccessfulEncodeDecode(SerialisationError),
}

Launcher Errors

Variants

CoreError

Error from safe_core. Boxed to hold a pointer instead of value so that this enum variant is not insanely bigger than others.

NfsError

Errors from safe_nfs

DnsError

Errors from safe_nfs

PathNotFound

Unable to find/traverse directory or file path

InvalidPath

Supplied path was invalid

PermissionDenied

Permission denied - e.g. permission to access SAFEDrive etc.

JsonParseError

Could not parse payload as a valid JSON

JsonDecodeError

Could not decode valid JSON into expected Structures probably because a mandatory field was missing or a field was wrongly named etc.

SpecificParseError

JSON non-conforming to the Launcher RFC and not covered by JsonDecodeError, e.g. things like invalid base64 formatting, unreasonable/unexpected indexing, ranges etc.

JsonEncodeError

Error encoding into Json String

LocalConfigAccessFailed

Unable to Read from or Write to a Local Config file.

Unexpected

Unexpected - Probably a Logic error

UnsuccessfulEncodeDecode

Could not serialise or deserialise data