Function safe_ffi::drop_client

pub extern fn drop_client(client_handle: *const c_void)

Discard and clean up the previously allocated client. Use this only if the client is obtained from one of the client obtainment functions in this crate (crate_account, log_in, create_unregistered_client). Using client_handle after a call to this functions is undefined behaviour.