Function safe_ffi::execute_for_content

pub extern fn execute_for_content(c_payload: *const c_char, c_size: *mut int32_t, c_capacity: *mut int32_t, c_result: *mut int32_t, ffi_handle: *const c_void) -> *const u8

General function that can be invoked for getting data as a resut for an operation. The function return a pointer to a U8 vecotr. The size of the U8 vector and its capacity is written to the out params c_size & c_capacity. The size and capcity would be required for droping the vector The result of the execution is returned in the c_result out param