Function safe_ffi::get_file_content_from_service_home_dir

pub extern fn get_file_content_from_service_home_dir(client_handle: *const c_void, c_long_name: *const c_char, c_service_name: *const c_char, c_file_name: *const c_char, c_content_buf: *mut uint8_t) -> int32_t

Get file content from service home directory The Name of the file is the final token in the given path. Eg., if given path = /a/b/c/d then d is interpreted as the file intended to be read. c_content_buf should be properly and sufficiently pre-allocated.