Crate safe_ffi

This crate provides FFI-bindings to the Client Modules (safe_client, safe_nfs, safe_dns) In the current implementation the allocations made by this crate are managed within the crate itself and is guaranteed that management of such allocations will not be pushed beyond the FFI boundary. This has a 2-fold outcome: firstly, the passing of data is done by filling of the allocations passed by the caller and is caller's responsibility to manage those. For this every function that fills an allocated memory also has a companion function to return the size of data which the caller can call to find out how much space needs to be allocated in the first place. Second and consequently, the caller does not have to bother calling functions within this crate which only serve to free resources allocated by the crate itself. This otherwise would be error prone and cumbersome. Instead the caller can use whatever idiom in his language to manage memory much more naturally and conveniently (eg., RAII idioms etc)

The only exception to the above rule is the obtainment of the client engine itself. The client engine is allocated and managed by the crate. This is necessary because it serves as a context to all operations provided by the crate. Hence the user will obtain the engine on calling any one of the functions to create it and must preserve it for all subsequent operations. When done, to release the resources, drop_client may be called.

Project github page

Macros

ffi_try!
This macro is intended to be used in all cases where we get Result and want to convert it to C friendly error code to inform about error conditions accross library boundaries

Functions

add_service
Add a new service to the existing (registered) Dns record

create_account
Create a registered client. This or any one of the other companion functions to get a client must be called before initiating any operation allowed by this crate. client_handle is a pointer to a pointer and must point to a valid pointer not junk, else the consequences are undefined.

create_file
Create a file. The Name of the file is the final token in the given path. Eg., if given path = /a/b/c/d then d is interpreted as the file intended to be created.

create_sub_directory
Create a subdirectory. The Name of the subdirectory is the final token in the given path. Eg., if given path = /a/b/c/d then d is interpreted as the subdirectory intended to be created.

create_unregistered_client
Create an unregistered client. This or any one of the other companion functions to get a client must be called before initiating any operation allowed by this crate.

drop_client
Discard and clean up the previously allocated client. Use this only if the client is obtained from one of the client obtainment functions in this crate (crate_account, log_in, create_unregistered_client). Using client_handle after a call to this functions is undefined behaviour.

get_file_content
Read a file. The Name of the file is the final token in the given path. Eg., if given path = /a/b/c/d then d is interpreted as the file intended to be read. c_content_buf should be properly and sufficiently pre-allocated.

get_file_content_from_service_home_dir
Get file content from service home directory The Name of the file is the final token in the given path. Eg., if given path = /a/b/c/d then d is interpreted as the file intended to be read. c_content_buf should be properly and sufficiently pre-allocated.

get_file_size
Get the size of the file. c_size should be properly and sufficiently pre-allocated. The Name of the file is the final token in the given path. Eg., if given path = /a/b/c/d then d is interpreted as the file intended to be read. c_size should be properly and sufficiently pre-allocated.

get_file_size_from_service_home_dir
Get file size from service home directory The Name of the file is the final token in the given path. Eg., if given path = /a/b/c/d then d is interpreted as the file intended to be read. c_content_size should be properly and sufficiently pre-allocated.

log_in
Log into a registered client. This or any one of the other companion functions to get a client must be called before initiating any operation allowed by this crate. client_handle is a pointer to a pointer and must point to a valid pointer not junk, else the consequences are undefined.

register_dns
Register Dns